Added new attribyte owner

+++ api/controllers/things/destroy-one-thing.js
@@ -15,13 +15,23 @@ module.exports = {
   },
 
 
-  exits: {},
+  exits: {
+    forbidden: {
+      description: 'The user making this request does\'t have the permissions to delete this thing.',
+      responseType:'forbidden' // как раньше res.forbidden(), сейчас это встроеная функция sails
+    }
+  },
 
 
-  fn: async function (inputs) {
-
+  fn: async function (inputs,exits) {
+    let thing = await Thing.findOne({
+      id: inputs.id
+    });
+    if (thing.owner !== this.req.me.id) {
+      throw 'forbidden';
+    }
     await Thing.destroy({id: inputs.id});
-    return;
+    return exits.success();
 
   }
